user sequence fix

+++ src/main/kotlin/eu/glatz/sifidraw/repository/UserRepository.kt
@@ -3,6 +3,6 @@ package eu.glatz.sifidraw.repository
 import eu.glatz.sifidraw.model.User
 import org.springframework.data.mongodb.repository.MongoRepository
 
-interface UserRepository : MongoRepository<User, Int> {
+interface UserRepository : MongoRepository<User, Long> {
     fun findByName(name: String): User?
 }
